Moved DB models to their own module
